pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "rcwa-app"
version = "0.1.0"
description = "Interactive, modular RCWA emissivity app (UI-first, compute-agnostic)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Josu M. Igartua" }]
keywords = ["RCWA", "emissivity", "optics", "periodic-structures", "streamlit", "xarray"]

# Runtime deps for the skeleton you ran
dependencies = [
  "numpy>=1.26",
  "xarray>=2024.3.0",
  "pydantic>=2.6",
  "plotly>=5.24",
  "streamlit>=1.38",
]

# Optional groups you can install with: pip install -e .[dev]  (and/or [test], [docs])
[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "black==24.8.0",
  "mypy>=1.10",
  "pytest>=8",
  "pytest-xdist>=3",
  "pytest-cov>=5",
  "hypothesis>=6",
  "build>=1.2",
  "twine>=5",
  "kaleido>=0.2.1", # enables Plotly static image export
  "pandas-stubs>=2.2.2.240807",
]

docs = ["mkdocs-material>=9"]

[project.scripts]
rcwa-app = "rcwa_app.__main__:main"

[project.urls]
Homepage = "https://example.com/rcwa-app"
Repository = "https://example.com/rcwa-app.git"

[tool.hatch.build.targets.wheel]
# Package only the library code; UI script stays in the repo for dev
packages = ["rcwa_app"]

# ---------------- Developer tool config ----------------
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# "tests/**" = ["D"]  # example, if you need it

[tool.black]
line-length = 100
target-version = ["py310","py311"]

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = true
pretty = true

# Per-module override: silence missing stubs for Plotly.
# Equivalent to INI-style `[mypy-plotly.*] ignore_missing_imports = True`.
[[tool.mypy.overrides]]
module = ["plotly.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
